Feat#52 - 회원가입 인증 구현, 비밀번호 초기화 기능 구현

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/auth/service/UserAuthService.java
@@ -11,7 +11,7 @@ public class UserAuthService {
     private final UserAuthRepository userAuthRepository;
 
     public UserAuth findById(Long userId) {
-        return userAuthRepository.findById(userId.intValue())
+        return userAuthRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
     }
 }
